refactor: 💡 import structure and resolver import
