+++ src/resolvers/test/test.schema.graphql
@@ -1,0 +1,5 @@
+type TestBlock @entity {
+  blockHeight: BigInt!,
+  createdAt: DateTime!,
+  poolId: Int!
+}
