+++ src/server-extension.ts
@@ -1,14 +1,25 @@
 import { Resolver } from 'type-graphql';
-import { TestBlock as TestBlockModel } from './generated/model/testBlock.model';
 import { entityOverTimeResolverFactory } from './resolvers/factory';
-
-import { TestBlock } from './resolvers/test.resolver';
+import { HistoricalBalance } from './resolvers/model/historicalBalance.model';
+import {
+    HistoricalBalanceLBP as HistoricalBalanceLBPModel,
+    HistoricalBalanceXYK as HistoricalBalanceXykModel,
+} from './generated/model';
 
 @Resolver()
-export class TestBlockOverTimeResolver extends entityOverTimeResolverFactory<TestBlock>(
-    TestBlock,
-    TestBlockModel,
-    'testBlocksOverTime',
-    'test_block',
+export class HistoricalBalancesLbpGroupedResolver extends entityOverTimeResolverFactory<HistoricalBalance>(
+    HistoricalBalance,
+    HistoricalBalanceLBPModel,
+    'historicalBalancesLbpGrouped',
+    'historical_balance_lbp',
+    'pool_id'
+) {}
+
+@Resolver()
+export class HistoricalBalancesXykGroupedResolver extends entityOverTimeResolverFactory<HistoricalBalance>(
+    HistoricalBalance,
+    HistoricalBalanceXykModel,
+    'historicalBalancesXykGrouped',
+    'historical_balance_xyk',
     'pool_id'
 ) {}
